all/examples/src/rabbit/requirements.txt
pika==1.2.0
amqpstorm==2.9.0
python-dotenv